Update `qdrant.cloud.cluster.backup.v1` to use enums (#51)
We were using string values with a list of possible options because we
thought enums weren't working well with the generated code for openapi.
However, it looks that is well supported now (enums can be passed as
string values) and we would like to use them.

We update only one API first to confirm it works successfully. After
that we will update all the other APIs.