Fix the script that generates the package.json (#181)
To generate the exports, the script was making the assumption that there
is only one file per folder (e.g
`qdrant/cloud/cluster/v1/cluster.proto`). That's true in almost all the
cases, but we have some folders that contain multiple files (e.g.
`qdrant/cloud/cluster/auth/v1/`). With this change we handle that case,
creating specific exports for the different files inside the folder. If
there is only one file we keep the current behavior.