Update validation for AccountInvite.user_role_ids (#217)
Update the validation to allow an empty list. This is necessary for
cases where a user does not have permission to list roles and therefore
cannot include specific role IDs in the invitation.

By default, the system attaches the Base role to all users. If the list
of roles in the invitation is empty, the user will simply receive the Base
role.